% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_lnc_threshold.R
\name{compute_lnc_threshold}
\alias{compute_lnc_threshold}
\title{Compute LNC Threshold}
\usage{
compute_lnc_threshold(d, k, n)
}
\arguments{
\item{d}{number of dimenions}

\item{k}{number of neighbors}

\item{n}{sample size}
}
\value{
log-scale value of alpha value will be on the interval (-inf,0)
}
\description{
Computes alpha tuning parameters using a fitted surrogate model. Uses the minimize MSE method of calulcating alpha
}
\examples{
compute_lnc_threshold(2,5,1000)

}
