% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_lnc_threshold.R
\name{estimate_mse}
\alias{estimate_mse}
\title{Estimate MSE of KNN Estimator}
\usage{
estimate_mse(k = 5, alpha = 0, d = 2, rho = 0, N = 1000, M = 100,
  cluster = NULL)
}
\arguments{
\item{alpha}{alpha if < 0 then log scale otherwise [0,1]}

\item{d}{dimension}

\item{rho}{MVN correlation (compound systemtric)}

\item{N}{sample size}

\item{M}{replications (outer loop)}

\item{cluster}{a parallel cluster object}
}
\description{
This function computes the MSE for a given set of tuning parameter alpha, dimension, size of neighborhood
}
