% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutual_information.R
\name{knn_mi_temp}
\alias{knn_mi_temp}
\title{k Nearest Neighbor Mutual Information Estimation}
\usage{
knn_mi_temp(data, splits, k, method = "KSG2")
}
\arguments{
\item{data}{numeric vector}

\item{splits}{lower bound}

\item{k}{upper bound}

\item{method}{bounded numeric vector}
}
\value{
bounded numeric vector
}
\description{
Here is an explaination of what the stuff does
}
