% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nearest_neighbors}
\alias{nearest_neighbors}
\title{compute nearest neighbors according to infinity norm}
\usage{
nearest_neighbors(data, k)
}
\arguments{
\item{data}{matrix. Each row is an observation.}

\item{k}{integer. Number of neighbors to observe.}
}
\value{
List of distances and indices of k nearest neighbors of each point in \code{data}.
}
\description{
compute nearest neighbors according to infinity norm
}
\examples{
set.seed(123)
X <- cbind(1:10)
nearest_neighbors(X,3)
}
