% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_lnc_threshold.R
\name{optimize_mse}
\alias{optimize_mse}
\title{Optimize MSE of alpha MI estimators}
\usage{
optimize_mse(rho, N, M, d, k, lower = -5, upper = -1e-10, num_iter = 25,
  init_size = 20, diagnostics = FALSE, cluster = NULL)
}
\arguments{
\item{cluster}{}
}
\description{
Optimize MSE of alpha MI estimators
}
