% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmvn.R
\name{rmvn}
\alias{rmvn}
\title{Random Multivariate Normal (MVN)}
\usage{
rmvn(n, mu = rep(0, d), cov_mat)
}
\arguments{
\item{n}{sample size}

\item{mu}{mean vector}

\item{cov_mat}{covariance matrix}
}
\value{
Matrix (n x d) of random MVN draws
}
\description{
Random Multivariate Normal (MVN)
}
\examples{
mnormt::rmnorm(200,0, matrix(c(1,0.9,0.9,1),2,2))
}
