% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_mi.R
\name{run_center_test}
\alias{run_center_test}
\title{test centering}
\usage{
run_center_test()
}
\arguments{
\item{x}{numeric vector}

\item{low}{lower bound}

\item{high}{upper bound}

\item{res}{bounded numeric vector}
}
\value{
bounded numeric vector
}
\description{
test centering
}
