% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_lnc_threshold.R
\name{compute_lnc_threshold}
\alias{compute_lnc_threshold}
\title{Title}
\usage{
compute_lnc_threshold(d, k, cached = TRUE)
}
\arguments{
\item{cached}{}
}
\value{
x
}
\description{
Title
}
\examples{
x+y = 2
}
