% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{knn_mi}
\alias{knn_mi}
\title{mutual information function}
\usage{
knn_mi(data, splits, options)
}
\arguments{
\item{x}{numeric vector}

\item{low}{lower bound}

\item{high}{upper bound}

\item{res}{bounded numeric vector}
}
\value{
bounded numeric vector
}
\description{
mutual information function
}
