% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_neighbors.R
\name{nearest_neighbors}
\alias{nearest_neighbors}
\title{compute nearest neighbors using infinity norm}
\usage{
nearest_neighbors(data, k)
}
\arguments{
\item{data}{matrix. Each row is an observation.}

\item{k}{integer. Number of neighbors to observe.}
}
\value{
List of distances and indices of k nearest neighbors of each point in \code{data}.
}
\description{
compute nearest neighbors using infinity norm
}
\examples{
X <- cbind(1:10)
nearest_neighbors(X,3)

set.seed(123)
X <- cbind(runif(100),runif(100))
plot(X,pch=20)
points(X[3,1],X[3,2],col='blue',pch=19, cex=1.5)
nn <- nearest_neighbors(X,5)
a = X[nn$nn_inds[3,-1],1]
b = X[nn$nn_inds[3,-1],2]
points(a,b,col='red',pch=19, cex=1.5)
}
