% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiceOptim_alpha_calibration.R,
%   R/compute_lnc_threshold.R
\name{optimize_mse}
\alias{optimize_mse}
\alias{optimize_mse}
\title{Optimize MSE of alpha MI estimators
Uses Gaussian process (GP) optimization to stochastically minimize the mean squared error of the LNC estimator}
\usage{
optimize_mse(rho, N, M, d, k, lower = -100, upper = -1e-10, num_iter = 10,
  init_size = 20, diagnostics = FALSE, cluster = NULL, verbose = FALSE)

optimize_mse(rho, N, M, d, k, lower = -100, upper = -1e-10, num_iter = 10,
  init_size = 20, diagnostics = FALSE, cluster = NULL, verbose = FALSE)
}
\arguments{
\item{rho}{Correlation of reference distribution}

\item{N}{Sample size}

\item{M}{Monte Carlo sample size}

\item{d}{Dimension}

\item{k}{Neighboorhood order}

\item{lower}{Lower bound of constrained optimization}

\item{upper}{Upper bound of constrained optimization}

\item{num_iter}{Number of iterations of GP optimization}

\item{init_size}{Number of initial evaluation for estimating GP}

\item{diagnostics}{FALSE feature not implimented yet}

\item{cluster}{A FORK cluster from parallel package}

\item{rho}{Correlation of reference distribution}

\item{N}{Sample size}

\item{M}{Monte Carlo sample size}

\item{d}{Dimension}

\item{k}{Neighboorhood order}

\item{num_iter}{Number of iterations of GP optimization}

\item{init_size}{Number of initial evaluation for estimating GP}

\item{cluster}{A FORK cluster from parallel package}
}
\description{
Optimize MSE of alpha MI estimators
Uses Gaussian process (GP) optimization to stochastically minimize the mean squared error of the LNC estimator

Optimize MSE of alpha MI estimators
Uses Gaussian process (GP) optimization to stochastically minimize the mean squared error of the LNC estimator
}
