% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmvn.R
\name{rmvn}
\alias{rmvn}
\title{Random Multivariate Normal MVN Generator}
\usage{
rmvn(n, mu = rep(0, d), cov_mat)
}
\arguments{
\item{n}{number of observations}

\item{mu}{d dimensional mean vector}

\item{cov_mat}{d x d covariance matrix}
}
\value{
Matrix (n x d) of random MVN draws
}
\description{
Random Multivariate Normal MVN Generator
}
\keyword{internal}
