% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmvn.R
\name{simulate_mvn}
\alias{simulate_mvn}
\title{Calibration Random Multivariate Normal}
\usage{
simulate_mvn(n, d, rho)
}
\arguments{
\item{n}{number of observations}

\item{d}{number of dimension}

\item{rho}{correlation of compound-symmetric covariance}
}
\value{
Matrix (n x d) of random MVN draws
}
\description{
Calibration Random Multivariate Normal
}
\keyword{internal}
